refactored value checks into separate file
