+++ src/Compilers/CSharp/Test/Emit/CodeGen/CodeGenInParametersTests.cs
@@ -98,5 +98,144 @@ class Program
                 Diagnostic(ErrorCode.ERR_BadParameterModifiers, "in").WithArguments("in", "ref").WithLocation(13, 37)
             );
         }
+
+        [Fact]
+        public void ReadonlyParamCannotAssign()
+        {
+            var text = @"
+class Program
+{
+    static void M(in int arg1, in (int Alice, int Bob) arg2)
+    {
+        arg1 = 1;
+        arg2.Alice = 2;
+
+        arg1 ++;
+        arg2.Alice --;
+
+        arg1 += 1;
+        arg2.Alice -= 2;
+    }
+}
+";
+
+            var comp = CreateCompilationWithMscorlib45(text, new[] { ValueTupleRef, SystemRuntimeFacadeRef });
+            comp.VerifyDiagnostics(
+                // (6,9): error CS8208: A readonly parameter cannot be assigned to
+                //         arg1 = 1;
+                Diagnostic(ErrorCode.ERR_AssignReadonlyParam, "arg1").WithLocation(6, 9),
+                // (7,9): error CS8209: Members of readonly parameter 'ref readonly (int Alice, int Bob)' cannot be assigned to
+                //         arg2.Alice = 2;
+                Diagnostic(ErrorCode.ERR_AssignReadonlyParam2, "arg2.Alice").WithArguments("ref readonly (int Alice, int Bob)").WithLocation(7, 9),
+                // (9,9): error CS8208: A readonly parameter cannot be assigned to
+                //         arg1 ++;
+                Diagnostic(ErrorCode.ERR_AssignReadonlyParam, "arg1").WithLocation(9, 9),
+                // (10,9): error CS8209: Members of readonly parameter 'ref readonly (int Alice, int Bob)' cannot be assigned to
+                //         arg2.Alice --;
+                Diagnostic(ErrorCode.ERR_AssignReadonlyParam2, "arg2.Alice").WithArguments("ref readonly (int Alice, int Bob)").WithLocation(10, 9),
+                // (12,9): error CS8208: A readonly parameter cannot be assigned to
+                //         arg1 += 1;
+                Diagnostic(ErrorCode.ERR_AssignReadonlyParam, "arg1").WithLocation(12, 9),
+                // (13,9): error CS8209: Members of readonly parameter 'ref readonly (int Alice, int Bob)' cannot be assigned to
+                //         arg2.Alice -= 2;
+                Diagnostic(ErrorCode.ERR_AssignReadonlyParam2, "arg2.Alice").WithArguments("ref readonly (int Alice, int Bob)").WithLocation(13, 9)
+            );
+        }
+
+        [Fact]
+        public void ReadonlyParamCannotAssignByref()
+        {
+            var text = @"
+class Program
+{
+    static void M(in int arg1, in (int Alice, int Bob) arg2)
+    {
+        ref var y = ref arg1;
+        ref int a = ref arg2.Alice;
+    }
+}
+";
+
+            var comp = CreateCompilationWithMscorlib45(text, new[] { ValueTupleRef, SystemRuntimeFacadeRef });
+            comp.VerifyDiagnostics(
+                // (6,25): error CS8206: A readonly parameter cannot be used as a ref or out value
+                //         ref var y = ref arg1;
+                Diagnostic(ErrorCode.ERR_RefReadonlyParam, "arg1").WithLocation(6, 25),
+                // (7,25): error CS8207: Members of readonly parameter 'ref readonly (int Alice, int Bob)' cannot be used as a ref or out value
+                //         ref int a = ref arg2.Alice;
+                Diagnostic(ErrorCode.ERR_RefReadonlyParam2, "arg2.Alice").WithArguments("ref readonly (int Alice, int Bob)").WithLocation(7, 25)
+            );
+        }
+
+        [Fact]
+        public void ReadonlyParamCannotReturnByOrdinaryRef()
+        {
+            var text = @"
+class Program
+{
+    static ref int M(in int arg1, in (int Alice, int Bob) arg2)
+    {
+        bool b = true;
+
+        if (b)
+        {
+            return ref arg1;
+        }
+        else
+        {
+            return ref arg2.Alice;
+        }
+    }
+}
+";
+
+            var comp = CreateCompilationWithMscorlib45(text, new[] { ValueTupleRef, SystemRuntimeFacadeRef });
+            comp.VerifyDiagnostics(
+                // (10,24): error CS8206: A readonly parameter cannot be used as a ref or out value
+                //             return ref arg1;
+                Diagnostic(ErrorCode.ERR_RefReadonlyParam, "arg1").WithLocation(10, 24),
+                // (14,24): error CS8207: Members of readonly parameter 'ref readonly (int Alice, int Bob)' cannot be used as a ref or out value
+                //             return ref arg2.Alice;
+                Diagnostic(ErrorCode.ERR_RefReadonlyParam2, "arg2.Alice").WithArguments("ref readonly (int Alice, int Bob)").WithLocation(14, 24)
+            );
+        }
+
+        [Fact]
+        public void ReadonlyParamCanReturnByRefReadonly()
+        {
+            var text = @"
+class Program
+{
+    static ref readonly int M(in int arg1, in (int Alice, int Bob) arg2)
+    {
+        bool b = true;
+
+        if (b)
+        {
+            return ref arg1;
+        }
+        else
+        {
+            return ref arg2.Alice;
+        }
+    }
+}
+";
+
+            var comp = CompileAndVerify(text, new[] { ValueTupleRef, SystemRuntimeFacadeRef }, parseOptions: TestOptions.Regular, verify: false);
+
+            comp.VerifyIL("Program.M", @"
+{
+  // Code size       12 (0xc)
+  .maxstack  1
+  IL_0000:  ldc.i4.1
+  IL_0001:  brfalse.s  IL_0005
+  IL_0003:  ldarg.0
+  IL_0004:  ret
+  IL_0005:  ldarg.1
+  IL_0006:  ldflda     ""int System.ValueTuple<int, int>.Item1""
+  IL_000b:  ret
+}");
+        }
     }
 }
